lisa: move lisa_m.cpp contents to lisa.cpp
